backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
gunicorn==22.0.0
python-dotenv==1.0.1
